+++ kernel/condvar.c
@@ -8,7 +8,7 @@
 #include <zephyr/kernel_structs.h>
 #include <zephyr/toolchain.h>
 #include <ksched.h>
-#include <zephyr/wait_q.h>
+#include <wait_q.h>
 #include <zephyr/syscall_handler.h>
 
 static struct k_spinlock lock;
