+++ kernel/kheap.c
@@ -5,11 +5,12 @@
  */
 
 #include <zephyr/kernel.h>
-#include <ksched.h>
-#include <zephyr/wait_q.h>
 #include <zephyr/init.h>
 #include <zephyr/linker/linker-defs.h>
 #include <zephyr/sys/iterable_sections.h>
+/* private kernel APIs */
+#include <ksched.h>
+#include <wait_q.h>
 
 void k_heap_init(struct k_heap *h, void *mem, size_t bytes)
 {
