+++ kernel/pipes.c
@@ -15,7 +15,7 @@
 
 #include <zephyr/toolchain.h>
 #include <ksched.h>
-#include <zephyr/wait_q.h>
+#include <wait_q.h>
 #include <zephyr/init.h>
 #include <zephyr/syscall_handler.h>
 #include <kernel_internal.h>
