+++ kernel/poll.c
@@ -17,7 +17,7 @@
 #include <zephyr/kernel.h>
 #include <zephyr/kernel_structs.h>
 #include <kernel_internal.h>
-#include <zephyr/wait_q.h>
+#include <wait_q.h>
 #include <ksched.h>
 #include <zephyr/syscall_handler.h>
 #include <zephyr/sys/dlist.h>
