+++ kernel/queue.c
@@ -15,7 +15,7 @@
 #include <zephyr/kernel_structs.h>
 
 #include <zephyr/toolchain.h>
-#include <zephyr/wait_q.h>
+#include <wait_q.h>
 #include <ksched.h>
 #include <zephyr/init.h>
 #include <zephyr/syscall_handler.h>
