+++ kernel/work.c
@@ -12,7 +12,7 @@
 
 #include <zephyr/kernel.h>
 #include <zephyr/kernel_structs.h>
-#include <zephyr/wait_q.h>
+#include <wait_q.h>
 #include <zephyr/spinlock.h>
 #include <errno.h>
 #include <ksched.h>
