+++ lib/os/p4wq.c
@@ -5,11 +5,12 @@
  */
 #include <zephyr/logging/log.h>
 #include <zephyr/sys/p4wq.h>
-#include <zephyr/wait_q.h>
 #include <zephyr/kernel.h>
-#include <ksched.h>
 #include <zephyr/init.h>
 #include <zephyr/sys/iterable_sections.h>
+/* private kernel APIs */
+#include <ksched.h>
+#include <wait_q.h>
 
 LOG_MODULE_REGISTER(p4wq, CONFIG_LOG_DEFAULT_LEVEL);
