+++ tests/kernel/events/event_api/CMakeLists.txt
@@ -6,3 +6,7 @@ project(events_api)
 
 FILE(GLOB app_sources src/*.c)
 target_sources(app PRIVATE ${app_sources})
+
+target_include_directories(app PRIVATE
+  ${ZEPHYR_BASE}/kernel/include
+  )
