+++ tests/kernel/events/event_api/src/test_event_apis.c
@@ -5,7 +5,7 @@
  */
 
 #include <zephyr/ztest.h>
-#include <zephyr/wait_q.h>
+#include <wait_q.h>
 
 #define DELAY          K_MSEC(50)
 #define SHORT_TIMEOUT  K_MSEC(100)
